multiple_bssid: make the AID space shared
There is a shared TIM IE in the transmitting BSSs beacon. For STAs connected to
the non-transmitting BSSs to be able to find their TIM bit we needs a shared
AID index. Also the AID0 used for mc/bc becomes AID<ifnum> so we need to adjust
the base offset.

Signed-off-by: John Crispin <john@phrozen.org>